+++ server/.gitignore
@@ -1,1 +1,1 @@
-data_storage
+data_storage/
