+++ server/settings/connectsettings.json
@@ -1,0 +1,7 @@
+{
+    "Connection Settings": [
+        {
+            "AES Encryption Password": "ThisIsAVerySecurePassword"
+        }
+    ]
+}
